backend/requirements.txt
fastapi
uvicorn
motor
pydantic
passlib[bcrypt]
python-dotenv
PyJWT
email-validator 